refactor(notes): address PR 1459 review feedback
- Extract the 3-dot actions into a shared <NoteActions> used by both the
  note card and the detail header; add a Delete option there.
- Export the note body as raw HTML (no tag stripping) and move the export
  helper to detail/utils.ts.
- Detail view takes full width (drop the max-w-200 cap).
- Resolve projectId via useProjectDetail instead of useParams.
- Use the long relative-time format in the detail header to match the card.
- Add a back button (ArrowLeft, mirroring Risks) before the avatar.
- Size dropdown option icons with `size-4 mr-2`.
