+++ frontend/packages/app/src/pages/project-details/tabs/notes/noteCard.tsx
@@ -1,17 +1,15 @@
 /**
  * External dependencies.
  */
-import { useNavigate, useParams, useSearchParams } from "react-router-dom";
-import { Avatar, Dropdown } from "@rtcamp/frappe-ui-react";
-import { DotHorizontal } from "@rtcamp/frappe-ui-react/icons";
+import { useSearchParams } from "react-router-dom";
+import { Avatar } from "@rtcamp/frappe-ui-react";
 
 /**
  * Internal dependencies.
  */
-import { ROUTES } from "@/lib/constant";
 import { formatRelativeTimeShort, stripTags } from "@/lib/utils";
 import { NOTE_PARAM } from "./constants";
-import { useNotes } from "./context";
+import { NoteActions } from "./noteActions";
 import type { Note } from "./types";
 
 type NoteCardProps = {
@@ -19,11 +17,7 @@ type NoteCardProps = {
 };
 
 export function NoteCard({ note }: NoteCardProps) {
-  const navigate = useNavigate();
   const [, setSearchParams] = useSearchParams();
-  const { deleteNote } = useNotes((s) => s.actions);
-  const isDeleting = useNotes((s) => s.state.isDeleting);
-  const { projectId = "" } = useParams<{ projectId: string }>();
   const excerpt = stripTags(note.description);
   const relativeDate = formatRelativeTimeShort(note.creation, new Date(), true);
   const authorHref = `/desk/user/${encodeURIComponent(note.owner)}`;
@@ -40,7 +34,10 @@ export function NoteCard({ note }: NoteCardProps) {
       tabIndex={0}
       onClick={openDetail}
       onKeyDown={(e) => {
-        if (e.target === e.currentTarget && (e.key === "Enter" || e.key === " ")) {
+        if (
+          e.target === e.currentTarget &&
+          (e.key === "Enter" || e.key === " ")
+        ) {
           e.preventDefault();
           openDetail();
         }
@@ -51,33 +48,7 @@ export function NoteCard({ note }: NoteCardProps) {
         <h3 className="flex-1 truncate text-lg font-medium text-ink-gray-8">
           {note.title}
         </h3>
-        <span onClick={(e) => e.stopPropagation()}>
-          <Dropdown
-            placement="right"
-            button={{
-              variant: "ghost",
-              icon: DotHorizontal,
-            }}
-            options={[
-              {
-                label: "Edit",
-                key: "edit",
-                disabled: isDeleting,
-                onClick: () =>
-                  navigate(
-                    `${ROUTES.project}/${projectId}/notes/${note.name}/edit`,
-                  ),
-              },
-              {
-                label: "Delete",
-                key: "delete",
-                theme: "red",
-                disabled: isDeleting,
-                onClick: () => deleteNote(note.name),
-              },
-            ]}
-          />
-        </span>
+        <NoteActions note={note} />
       </div>
       <p className="flex-1 line-clamp-7 whitespace-pre-wrap px-3.5 pt-2 text-base leading-6 text-ink-gray-5">
         {excerpt}
